+++ vedo/settings.py
@@ -41,6 +41,10 @@ General settings.
     # Allow to interact with scene during interactor.Start() execution
     allowInteraction = True
 
+    # Set up default mouse and keyboard functionalities
+    enableDefaultMouseCallbacks = True
+    enableDefaultKeyboardCallbacks = True
+
     # If False, when multiple renderers are present do not render each one for separate
     #  but do it just once at the end (when interactive() is called)
     immediateRendering = True
@@ -91,8 +95,8 @@ General settings.
     maxNumberOfPeels= 8     # maximum number of rendering passes
     occlusionRatio  = 0.0   # occlusion ratio, 0 = exact image.
 
-    # Turn on/off nvidia FXAA anti-aliasing, if supported.
-    useFXAA = False         # either True or False. None sets the system default
+    # Turn on/off nvidia FXAA post-process anti-aliasing, if supported.
+    useFXAA = False         # either True or False
 
     # By default, the depth buffer is reset for each renderer. If True, use the existing depth buffer
     preserveDepthBuffer = False
@@ -142,6 +146,10 @@ import os, vtk
 __all__ = ['datadir', 'dataurl', 'embedWindow']
 
 
+vtk_version = [ vtk.vtkVersion().GetVTKMajorVersion(),
+                vtk.vtkVersion().GetVTKMinorVersion(),
+                vtk.vtkVersion().GetVTKBuildVersion() ]
+
 ####################################################################################
 # Axes titles
 xtitle = 'x'
@@ -168,6 +176,10 @@ interactorStyle = None
 
 # Allow to interact with scene during interactor.Start() execution
 allowInteraction = True
+
+# Set up default mouse and keyboard functionalities
+enableDefaultMouseCallbacks = True
+enableDefaultKeyboardCallbacks = True
 
 # When multiple renderers are present do not render each one for separate.
 # but do it just once at the end (when interactive() is called)
@@ -208,18 +220,16 @@ twoSidedLighting = True
 
 # Turn on/off rendering of translucent material with depth peeling technique.
 #https://lorensen.github.io/VTKExamples/site/Cxx/Visualization/CorrectlyRenderTranslucentGeometry
-
 useDepthPeeling = False
 multiSamples = 8
-if vtk.vtkVersion().GetVTKMajorVersion() >= 9:
+if vtk_version[0] >= 9:
     useDepthPeeling = True
+maxNumberOfPeels= 8
 occlusionRatio  = 0.1
-maxNumberOfPeels= 16
-occlusionRatio  = 0.0
 alphaBitPlanes  = True
 
 # Turn on/off nvidia FXAA anti-aliasing, if supported.
-useFXAA         = False  # either True or False. None sets the system default
+useFXAA         = False  # either True or False
 
 # By default, the depth buffer is reset for each renderer. If true, use the existing depth buffer
 preserveDepthBuffer = False
@@ -473,6 +483,14 @@ font_parameters = dict(
                         mono = False,
                         fscale = 0.8,
                         hspacing = 0.9,
+                        lspacing = 0.225,
+                        dotsep = "~^.~ ",
+                        islocal = False,
+                        ),
+        Komika= dict(
+                        mono = False,
+                        fscale = 0.7,
+                        hspacing = 0.75,
                         lspacing = 0.225,
                         dotsep = "~^.~ ",
                         islocal = False,